+++ doc/whats_new.rst
@@ -19,6 +19,10 @@ Current
 Changelog
 ~~~~~~~~~
 
+- Add new tutorial on :ref:`sphx_glr_auto_tutorials_plot_sleep.py` by `Alex Gramfort`_, `Stanislas Chambon`_ and `Joan Massich`_
+
+- Add data fetchers for polysomnography (PSG) recordings from Physionet (:func:`mne.datasets.sleep_physionet.age.fetch_data` and :func:`mne.datasets.sleep_physionet.temazepam.fetch_data`) by `Alex Gramfort`_ and `Joan Massich`_
+
 - Add support for indexing, slicing, and iterating :class:`mne.Annotations` by `Joan Massich`_
 
 - :meth:`mne.io.Raw.plot` now uses the lesser of ``n_channels`` and ``raw.ch_names``, by `Joan Massich`_
@@ -36,6 +40,7 @@ Changelog
 - Add :func:`mne.labels_to_stc` to facilitate working with label data, by `Eric Larson`_
 
 - Add ``overlap`` argument to :func:`mne.make_fixed_length_events` by `Eric Larson`_
+
 - Add 448-labels subdivided aparc cortical parcellation by `Denis Engemann`_ and `Sheraz Khan`_
 
 - Add keyboard shortcuts to nativate volume source estimates in time using (shift+)left/right arrow keys by `Mainak Jas`_
@@ -3134,3 +3139,5 @@ of commits):
 .. _David Haslacher: https://github.com/davidhaslacher
 
 .. _Larry Eisenman:  https://github.com/lneisenman
+
+.. _Stanislas Chambon: https://github.com/Slasnista